generic version of doDeath!
